Add files via upload
Airbnb, an online platform for peer-to-peer short-term accommodation rental, is growing with a spectacular speed in most countries of the world. This creates a great regulatory challenge, as empirical evidence suggest that Airbnb may have a significant impact on the traditional hotel industry and on the housing market. The aim of this analysis is to find the price of the new hosts based on their listings of Airbnb to traditional services and to examine the standard between the different business models. The empirical analysis is based on a unique dataset of scraped data on the listings of Airbnb in City of Barcelona, Spain. 

Objective:
The objective of this project is to do a research and build models for Airbnb Price Prediction. The scope of the project is that the host might not know how much they should price their listing and this affects the company's revenue. By analyzing and considering the aspects of the listing types, location, reviews, price near neighbourhood and the amenities in their listings, we guide the new hosts to set an optimal pricing for their listings. There are some cases where the hosts are new to the locality and don’t know the optimum price of the listing and end up setting up some random prices. Hence we alert the hosts to change the price accordingly.